+++ helper_test.go
@@ -9,6 +9,8 @@ import (
 	"github.com/goyek/goyek/v2"
 )
 
+var noDeps goyek.Deps
+
 func assertTrue(tb testing.TB, got bool, msg string) {
 	tb.Helper()
 	if got {